backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.354
langchain-community==0.0.10
langchain-openai==0.0.2
langgraph==0.0.40
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
duckduckgo-search==3.9.11
arxiv==2.1.0
pydantic==2.5.0
nest-asyncio==1.5.8